backend/requirements.txt
# Django Core
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3

# Database (psycopg3 para Python 3.13+)
psycopg[binary]==3.2.3

# Authentication
djangorestframework-simplejwt==5.4.0
PyJWT==2.10.1

# Excel handling
openpyxl==3.1.5
et-xmlfile==2.0.0

# PDF generation
reportlab==4.2.5
pillow==11.1.0

# AFIP Integration (opcional - solo si vas a usar facturación real)
# pyafipws==2024.1.0

# Tickets térmicos (temporal: comentado hasta que soporte Python 3.13)
# python-escpos==3.0

# Utilities
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3  # Clover API (y otras integraciones HTTP)

# Development
ipython==8.31.0
django-debug-toolbar==4.4.6

# Testing
pytest==8.3.4
pytest-django==4.9.0
factory-boy==3.3.1

# Code quality
black==24.10.0
flake8==7.1.1
isort==5.13.2

# Production
gunicorn==23.0.0
whitenoise==6.8.2
